third_party/WebKit/LayoutTests/fast/css/usecounter-pseudo-webkit-any-link.html

<!DOCTYPE html>
<html>
  <head>
    <title>Use counter test for :-webkit-any-link</title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <style id="style"></style>
    <a href="http://google.com" id="link">Link</a>
    <script>
      'use strict';
      test(() => {
        let CSSSelectorPseudoWebkitAnyLink = 2260; // From UseCounter.h
        let isCounted = () => internals.isUseCounted(document, CSSSelectorPseudoWebkitAnyLink);
        var div = document.createElement('div');
        getComputedStyle(link);
        assert_false(isCounted(), ':-webkit-any-link should not be counted');
        var style = document.getElementById('style');
        style.textContent = ':-webkit-any-link { color: blue; }';
        assert_equals(getComputedStyle(link).color, 'rgb(0, 0, 255)');
        assert_true(isCounted(), ':-webkit-any-link should be counted');
      }, ':-webkit-any-link counter is working');
    </script>
  </head>
<html>
